requirements.txt
numpy
matplotlib
opencv-python
tensorflow
scikit-learn
pillow
random
joblib
tkinter
customtkinter